refactor: テーブルページのファイル名を適切な名前に修正
